+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Please refer to the [official documentation](https://wilfred.readthedocs.io/en/latest/development/) for more information about the CHANGELOG and releases.
 
+## next-release
+
+### Added
+
+* Check for new releases against GitHub when running `wilfred --version`.
+
 ## v0.2.0 (released on 2020-01-18)
 
 ### Added